modal_deployment/requirements.txt
modal>=0.55.0
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
matplotlib>=3.4.0
requests>=2.25.0